+++ Code/Launcher/Launcher.cpp
@@ -102,7 +102,11 @@ void Launcher::LoadEngine()
 	if (!m_params.isDedicatedServer)
 	{
 		m_CryAction.Load("CryAction.dll", DLL::NO_UNLOAD);
-		m_CryRenderD3D10.Load("CryRenderD3D10.dll", DLL::NO_UNLOAD);
+
+		if (!CmdLine::HasArg("-dx9"))
+		{
+			m_CryRenderD3D10.Load("CryRenderD3D10.dll", DLL::NO_UNLOAD);
+		}
 	}
 }
 
@@ -171,7 +175,7 @@ void Launcher::PatchEngine_CrySystem()
 
 void Launcher::PatchEngine_CryRenderD3D10()
 {
-	if (!m_params.isDedicatedServer)
+	if (m_CryRenderD3D10.IsLoaded())
 	{
 		void *pCryRenderD3D10 = m_CryRenderD3D10.GetHandle();